+++ R/comparisons.R
@@ -46,6 +46,7 @@
 #'   - Numeric variables:
 #'     * Numeric of length 1: Contrast for a gap of `x`, computed at the observed value plus and minus `x / 2`. For example, estimating a `+1` contrast compares adjusted predictions when the regressor is equal to its observed value minus 0.5 and its observed value plus 0.5.
 #'     * Numeric vector of length 2: Contrast between the 2nd element and the 1st element of the `x` vector.
+#'     * Function which accepts a numeric vector and returns a data frame with two columsn of "low" and "high" values to compare. See examples.
 #'     * "iqr": Contrast across the interquartile range of the regressor.
 #'     * "sd": Contrast across one standard deviation around the regressor mean.
 #'     * "2sd": Contrast across two standard deviations around the regressor mean.
@@ -121,6 +122,13 @@
 #' comparisons(mod, variables = list(hp = "iqr")) %>% tidy()
 #' comparisons(mod, variables = list(hp = "sd")) %>% tidy()
 #' comparisons(mod, variables = list(hp = "minmax")) %>% tidy()
+#' 
+#' # using a function to specify a custom difference in one regressor
+#' dat <- mtcars
+#' dat$new_hp <- 49 * (dat$hp - min(dat$hp)) / (max(dat$hp) - min(dat$hp)) + 1
+#' modlog <- lm(mpg ~ log(new_hp) + factor(cyl), data = dat)
+#' fdiff <- \(x) data.frame(x, x + 10)
+#' comparisons(modlog, variables = list(new_hp = fdiff)) %>% summary()
 #'
 #' # Adjusted Risk Ratio: see the contrasts vignette
 #' mod <- glm(vs ~ mpg, data = mtcars, family = binomial)