+++ R/sanitize_variables.R
@@ -169,7 +169,8 @@ sanitize_variables <- function(variables,
                     # Eventually it would be nice to consolidate, but that's a lot of work.
                     valid_str <- c("iqr", "minmax", "sd", "2sd")
                     flag <- isTRUE(checkmate::check_numeric(predictors[[v]], min.len = 1, max.len = 2)) ||
-                            isTRUE(checkmate::check_choice(predictors[[v]], choices = valid_str))
+                            isTRUE(checkmate::check_choice(predictors[[v]], choices = valid_str)) ||
+                            isTRUE(checkmate::check_function(predictors[[v]]))
                     if (!isTRUE(flag)) {
                         msg <- "The %s element of the `variables` argument is invalid."
                         msg <- sprintf(msg, v)